requirements.txt
astroid==2.3.3
autopep8==1.4.4
cffi==1.13.2
Click==7.0
cryptography==2.8
entrypoints==0.3
flake8==3.7.9
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2==2.8.4
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
PyJWT==1.7.1
six==1.13.0
SQLAlchemy==1.3.11
typed-ast==1.4.0
Werkzeug==0.16.0
wrapt==1.11.2
